+++ Events.php
@@ -7,7 +7,7 @@
  */
 
 /**
- *  Copyright (c) Ascensio System SIA 2023. All rights reserved.
+ *  Copyright (c) Ascensio System SIA 2024. All rights reserved.
  *  http://www.onlyoffice.com
  */
 
@@ -24,7 +24,6 @@ use humhub\modules\onlyoffice\permissions\CanUseOnlyOffice;
  */
 class Events
 {
-
     public static function onFileHandlerCollection($event)
     {
         if (!Yii::$app->user->can(CanUseOnlyOffice::class)) {
@@ -43,20 +42,28 @@ class Events
         $module = Yii::$app->getModule('onlyoffice');
         $file = $event->sender->file;
 
-        if ($module->getDocumentType($file) !== null) {
-            $canEdit = $collection->type == FileHandlerCollection::TYPE_EDIT && $module->canEdit($file);
-            $canConvert = $collection->type == FileHandlerCollection::TYPE_EDIT && $module->canConvert($file);
-            $canView = $collection->type == FileHandlerCollection::TYPE_VIEW && $module->canView($file);
+        if ($module->getDocumentType($file) === null) {
+            return;
+        }
 
-            if ($canEdit) {
-                $collection->register(new filehandler\EditFileHandler());
-            }
-            if ($canConvert) {
-                $collection->register(new filehandler\ConvertFileHandler());
-            }
-            if ($canView) {
-                $collection->register(new filehandler\ViewFileHandler());
-            }
+        $isOnlyofficeForm = $module->isOnlyofficeForm($file);
+
+        $canEdit = $collection->type == FileHandlerCollection::TYPE_EDIT && $module->canEdit($file);
+        $canConvert = $collection->type == FileHandlerCollection::TYPE_EDIT && $module->canConvert($file);
+        $canView = $collection->type == FileHandlerCollection::TYPE_VIEW && $module->canView($file);
+
+        if ($canEdit) {
+            $collection->register(new filehandler\EditFileHandler());
+        }
+        if ($canConvert) {
+            $collection->register(new filehandler\ConvertFileHandler());
+        }
+        if ($canView && $isOnlyofficeForm) {
+            $collection->register(new filehandler\FillingFileHandler());
+            return;
+        }
+        if ($canView) {
+            $collection->register(new filehandler\ViewFileHandler());
         }
     }