+++ controllers/ApiController.php
@@ -7,7 +7,7 @@
  */
 
 /**
- *  Copyright (c) Ascensio System SIA 2023. All rights reserved.
+ *  Copyright (c) Ascensio System SIA 2024. All rights reserved.
  *  http://www.onlyoffice.com
  */
 
@@ -17,7 +17,7 @@ use Yii;
 use Exception;
 use yii\helpers\Url;
 use humhub\components\Controller;
-use \humhub\components\Module;
+use humhub\components\Module;
 use humhub\modules\file\models\File;
 use humhub\modules\file\libs\FileHelper;
 use humhub\modules\user\models\User;
@@ -51,6 +51,12 @@ class ApiController extends Controller
         $url = Yii::$app->request->post('url');
         $filename = Yii::$app->request->post('name');
 
+        if (parse_url($url, PHP_URL_HOST) !== parse_url($this->module->getServerUrl(), PHP_URL_HOST)) {
+            throw new Exception("Incorrect domain in file url");
+        }
+
+        $url = $this->module->replaceDocumentServerUrlToInternal($url);
+
         $response = $this->module->request($url);
 
         $newContent = $response->getContent();
@@ -59,7 +65,9 @@ class ApiController extends Controller
         $file = new File();
         $file->file_name = $filename;
         $file->size = mb_strlen($newContent, '8bit');
-        $file->mime_type = isset($this->module->mimes[$fileExt]) ? $this->module->mimes[$fileExt] : 'application/octet-stream';
+        $file->mime_type = isset($this->module->formats()->mimes[$fileExt])
+            ? $this->module->formats()->mimes[$fileExt]
+            : 'application/octet-stream';
         $file->save();
         $file->getStore()->setContent($newContent);
 
@@ -75,8 +83,13 @@ class ApiController extends Controller
         $users = User::find()->all();
 
         foreach ($users as $user) {
-            if ($user->id != $curUser->id && $user->profile->firstname != null && $user->profile->lastname != null && $user->email != null) {
-                array_push($usersForMentions,[
+            if (
+                $user->id != $curUser->id &&
+                $user->profile->firstname != null &&
+                $user->profile->lastname != null &&
+                $user->email != null
+            ) {
+                array_push($usersForMentions, [
                     "name" => $user->profile->firstname . " " . $user->profile->lastname,
                     "email" => $user->email
                 ]);
@@ -94,13 +107,13 @@ class ApiController extends Controller
 
         $file = File::findOne(['onlyoffice_key' => $doc_key]);
         $url = Url::to(['/onlyoffice/open', 'guid' => $file->guid, 'mode' => 'view']);
-        
+
         return $this->asJson([
             'url' => $url
         ]);
     }
 
-     public function actionSendNotify()
+    public function actionSendNotify()
     {
         $emails = Yii::$app->request->post('emails');
         $message = Yii::$app->request->post('comment');
@@ -114,9 +127,9 @@ class ApiController extends Controller
 
         $mention = Mention::generateMention($file, $message, $anchor);
 
-        try{
+        try {
             Notify::instance()->from($originator)->about($mention)->sendBulk($users);
-        } catch(Exception $exception) {
+        } catch (Exception $exception) {
             throw new Exception("Mention error.");
         }
 
@@ -135,19 +148,19 @@ class ApiController extends Controller
         $ext = Yii::$app->request->post('ext');
 
         $file = File::findOne(['onlyoffice_key' => $key]);
-        
+
         $owner = User::findOne($file->created_by);
         $containerRecord = ContentContainer::findOne(['id' => $owner->contentcontainer_id]);
         $container = $containerRecord->getPolymorphicRelation();
         $canRename = $container->can(ManageContent::class);
-        if(!$canRename) {
+        if (!$canRename) {
             throw new \Exception('Permission denied');
         }
 
         $arrayName = explode(".", $newFileName);
         $curExt = end($arrayName);
 
-        if($ext !== $curExt) {
+        if ($ext !== $curExt) {
             $newFileName .= "." . $ext;
         }
 
@@ -156,13 +169,13 @@ class ApiController extends Controller
         $meta = [
             "c" => "meta",
             "key" => $key,
-            "meta"=> [
+            "meta" => [
                 "title" => $newFileName
             ]
         ];
         $response = $this->module->commandService($meta);
 
-        if($response['error'] !== 0){
+        if ($response['error'] !== 0) {
             throw new \Exception('Error from command Service: ' . $response['error']);
         }