+++ controllers/CreateController.php
@@ -7,7 +7,7 @@
  */
 
 /**
- *  Copyright (c) Ascensio System SIA 2023. All rights reserved.
+ *  Copyright (c) Ascensio System SIA 2024. All rights reserved.
  *  http://www.onlyoffice.com
  */
 
@@ -22,11 +22,10 @@ use humhub\modules\onlyoffice\Module;
 
 class CreateController extends \humhub\components\Controller
 {
-
     /**
      * @inheritdoc
      */
-    public function getAccessRules()
+    protected function getAccessRules()
     {
         return [
             [ControllerAccess::RULE_PERMISSION => [CanUseOnlyOffice::class]],
@@ -45,20 +44,19 @@ class CreateController extends \humhub\components\Controller
         $model->extension = Yii::$app->request->get('extension');
 
         if ($model->load(Yii::$app->request->post())) {
-
             $file = $model->save();
 
             if ($file !== false) {
                 return $this->asJson([
-                            'success' => true,
-                            'file' => FileHelper::getFileInfos($file),
-                            'openFlag' => (boolean) $model->openFlag,
-                            'openUrl' => Url::to(['/onlyoffice/open', 'guid' => $file->guid, 'mode' => Module::OPEN_MODE_EDIT])
+                    'success' => true,
+                    'file' => FileHelper::getFileInfos($file),
+                    'openFlag' => (bool) $model->openFlag,
+                    'openUrl' => Url::to(['/onlyoffice/open', 'guid' => $file->guid, 'mode' => Module::OPEN_MODE_EDIT])
                 ]);
             } else {
                 return $this->asJson([
-                            'success' => false,
-                            'output' => $this->renderAjax('document', ['model' => $model])
+                    'success' => false,
+                    'output' => $this->renderAjax('document', ['model' => $model])
                 ]);
             }
         }
@@ -76,5 +74,4 @@ class CreateController extends \humhub\components\Controller
 
         return null;
     }
-
 }