+++ filehandler/EditFileHandler.php
@@ -7,7 +7,7 @@
  */
 
 /**
- *  Copyright (c) Ascensio System SIA 2023. All rights reserved.
+ *  Copyright (c) Ascensio System SIA 2024. All rights reserved.
  *  http://www.onlyoffice.com
  */
 
@@ -25,7 +25,6 @@ use humhub\modules\file\handler\BaseFileHandler;
  */
 class EditFileHandler extends BaseFileHandler
 {
-
     /**
      * @inheritdoc
      */
@@ -33,17 +32,16 @@ class EditFileHandler extends BaseFileHandler
     {
         $attributes = [
             'label' => Yii::t('OnlyofficeModule.base', 'Edit document'),
-            'data-action-url' => Url::to(['/onlyoffice/open', 'guid' => $this->file->guid, 'mode' => Module::OPEN_MODE_EDIT]),
+            'data-action-url' => Url::to([
+                '/onlyoffice/open',
+                'guid' => $this->file->guid,
+                'mode' => Module::OPEN_MODE_EDIT
+            ]),
             'data-action-click' => 'ui.modal.load',
             'data-modal-id' => 'onlyoffice-modal',
             'data-modal-close' => ''
         ];
 
-        if (pathinfo($this->file->file_name, PATHINFO_EXTENSION) === 'oform') {
-            $attributes['label'] = Yii::t('OnlyofficeModule.base', 'Fill in form in ONLYOFFICE');
-        }
-
         return $attributes;
     }
-
 }