+++ views/open/index.php
@@ -1,18 +1,21 @@
 <?php
 
 /**
- *  Copyright (c) Ascensio System SIA 2023. All rights reserved.
+ *  Copyright (c) Ascensio System SIA 2024. All rights reserved.
  *  http://www.onlyoffice.com
  */
 
 use yii\web\View;
+
 ?>
+
 <div class="modal-dialog animated fadeIn" style="width:96%">
     <div class="modal-content onlyofficeModal" style="background-color:transparent;">
         <?=
         \humhub\modules\onlyoffice\widgets\EditorWidget::widget([
             'file' => $file,
             'mode' => $mode,
+            'restrict' => $restrict,
             'anchor' => $anchor
         ]);
         ?>
@@ -32,5 +35,4 @@ use yii\web\View;
         function setSize() {
             $(".onlyofficeModal").css("height", window.innerHeight - 110 + "px");
         }
-    ');
-?>
+    '); ?>