+++ widgets/views/editor.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- *  Copyright (c) Ascensio System SIA 2023. All rights reserved.
+ *  Copyright (c) Ascensio System SIA 2024. All rights reserved.
  *  http://www.onlyoffice.com
  */
 
@@ -20,16 +20,42 @@ if ($documentType === Module::DOCUMENT_TYPE_SPREADSHEET) {
     $headerBackgroundColor = '#5A7DC9';
 } elseif ($documentType === Module::DOCUMENT_TYPE_PRESENTATION) {
     $headerBackgroundColor = '#DD682B';
+} elseif ($documentType === Module::DOCUMENT_TYPE_PDF) {
+    $headerBackgroundColor = '#D45757';
 }
 ?>
 
 <?= Html::beginTag('div', $options) ?>
-<div style = "height:50px; border-radius: 8px 8px 0px 0px; background-color:<?= $headerBackgroundColor; ?>; padding-top:7px; padding-right:7px">
+<div style="height:50px;
+    border-radius: 8px 8px 0px 0px;
+    background-color:<?= $headerBackgroundColor; ?>;
+    padding-top:7px;
+    padding-right:7px">
     <div class = "pull-right">
-        <?php if ($mode === Module::OPEN_MODE_EDIT && !Yii::$app->user->isGuest): ?>
-            <?= humhub\libs\Html::a(Yii::t('OnlyofficeModule.base', 'Share'), '#', ['class' => 'btn btn btn-default', 'data-action-click' => 'share', 'data-action-block' => 'sync', 'data-action-url' => Url::to(['/onlyoffice/share', 'guid' => $file->guid, 'mode' => $mode])]); ?>
+        <?php if ($mode === Module::OPEN_MODE_EDIT && !Yii::$app->user->isGuest) : ?>
+            <?= humhub\libs\Html::a(
+                Yii::t('OnlyofficeModule.base', 'Share'),
+                '#',
+                ['class' => 'btn btn btn-default',
+                    'data-action-click' => 'share',
+                    'data-action-block' => 'sync',
+                    'data-action-url' => Url::to([
+                        '/onlyoffice/share',
+                        'guid' => $file->guid,
+                        'mode' => $mode
+                    ])
+                ]
+            ); ?>
         <?php endif; ?>
-        <?= humhub\libs\Html::a(Yii::t('OnlyofficeModule.base', 'Close'), '#', ['class' => 'btn btn btn-default', 'data-ui-loader' => '', 'data-action-click' => 'close', 'data-action-block' => 'manual']); ?>
+        <?= humhub\libs\Html::a(
+            Yii::t('OnlyofficeModule.base', 'Close'),
+            '#',
+            ['class' => 'btn btn btn-default',
+                'data-ui-loader' => '',
+                'data-action-click' => 'close',
+                'data-action-block' => 'manual'
+            ]
+        ); ?>
     </div>
 </div>
 <div id="iframeContainer"></div>